del if field less then float number

+++ src/app/menu/menu_del_data/menu_del_less_value_field/del_if_less_depth/del_if_less_depth.c
@@ -2,9 +2,9 @@
 
 void del_if_less_depth(struct reservoirs_of_the_country *array, int length)
 {
-    int number;
+    float number;
     printf("Удалить поля 'Глубин', которые меньше, чем: ");
-    scanf("%d", &number);
+    scanf("%f", &number);
 
     for (int i = 0; i <= length; i++)
     {
