+++ src/app/menu/menu_del_data/menu_del_less_value_field/del_if_less_width/del_if_less_width.c
@@ -2,9 +2,9 @@
 
 void del_if_less_width(struct reservoirs_of_the_country *array, int length)
 {
-    int number;
+    float number;
     printf("Удалить поля 'Ширин', которые меньше, чем: ");
-    scanf("%d", &number);
+    scanf("%f", &number);
 
     for (int i = 0; i <= length; i++)
     {
